Move files to Dynamic subdirectory and adapt namespace
